# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RADapp AI Radiology Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Radiology report operations
  - name: chat
    description: AI chat assistant operations
  - name: dashboard
    description: Dashboard and stats operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports
      description: Returns all radiology reports for the user
      parameters:
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a radiology report
      parameters:
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created report with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /reports/recent:
    get:
      operationId: getRecentReports
      tags: [reports]
      summary: Get recent reports
      description: Returns most recent 5 reports for the user
      parameters:
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Recent reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /reports/{id}/share:
    post:
      operationId: shareReport
      tags: [reports]
      summary: Generate a shareable link token for a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Share token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareToken"

  /reports/shared/{token}:
    get:
      operationId: getSharedReport
      tags: [reports]
      summary: Get a report by share token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      description: Returns aggregate stats for the dashboard
      parameters:
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat history
      parameters:
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
        - name: reportId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Chat message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to AI assistant
      parameters:
        - name: x-user-id
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Report:
      type: object
      required: [id, title, createdAt, urgency, status]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        originalText:
          type: ["string", "null"]
        simplifiedExplanation:
          type: ["string", "null"]
        urgency:
          type: string
          enum: [low, moderate, high]
        status:
          type: string
          enum: [pending, processing, complete]
        recommendedNextSteps:
          type: ["string", "null"]
        medicalTermsBreakdown:
          type: ["string", "null"]
        reportType:
          type: ["string", "null"]
        bodyPart:
          type: ["string", "null"]
        shareToken:
          type: ["string", "null"]
        createdAt:
          type: string

    ReportInput:
      type: object
      required: [title, originalText]
      properties:
        title:
          type: string
          minLength: 1
        originalText:
          type: string
        reportType:
          type: string
        bodyPart:
          type: string

    ShareToken:
      type: object
      required: [token, url]
      properties:
        token:
          type: string
        url:
          type: string

    DashboardStats:
      type: object
      required: [totalReports, urgencyBreakdown, recentActivity]
      properties:
        totalReports:
          type: integer
        urgencyBreakdown:
          type: object
          required: [low, moderate, high]
          properties:
            low:
              type: integer
            moderate:
              type: integer
            high:
              type: integer
        recentActivity:
          type: integer
        averageProcessingTime:
          type: ["number", "null"]

    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        reportId:
          type: ["integer", "null"]
        createdAt:
          type: string

    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        reportId:
          type: ["integer", "null"]